feat(BACKEND): :triangular_flag_on_post: Update Backend API Calc

+++ api/database/migrations/2025_10_07_190902_add_discount_fields_to_payments_table.php
@@ -14,10 +14,10 @@ return new class extends Migration
     public function up()
     {
         Schema::table('payments', function (Blueprint $table) {
-            $table->integer('discount_applied')->nullable()->after('status'); // desconto total aplicado (pro-rata + créditos)
-            $table->integer('prorated_old')->nullable()->after('discount_applied'); // valor pro-rata do plano antigo
-            $table->integer('prorated_new')->nullable()->after('prorated_old'); // valor pro-rata do plano novo
-            $table->integer('applied_credits')->nullable()->after('prorated_new'); // créditos aplicados do saldo
+            $table->decimal('discount_applied', 13, 2)->nullable()->after('status'); // desconto total aplicado (pro-rata + créditos) EM REAIS
+            $table->decimal('prorated_old', 13, 2)->nullable()->after('discount_applied'); // valor pro-rata do plano antigo EM REAIS
+            $table->decimal('prorated_new', 13, 2)->nullable()->after('prorated_old'); // valor pro-rata do plano novo EM REAIS
+            $table->decimal('applied_credits', 13, 2)->nullable()->after('prorated_new'); // créditos aplicados do saldo EM REAIS
         });
     }
